Сделал все контроллеры и обработчики исключений

+++ src/main/java/org/example/taskmanagmentsystem/service/TaskService.java
@@ -1,18 +1,29 @@
 package org.example.taskmanagmentsystem.service;
 
 import jakarta.persistence.EntityNotFoundException;
+import lombok.extern.log4j.Log4j2;
 import org.example.taskmanagmentsystem.dto.TaskDTO;
 import org.example.taskmanagmentsystem.entity.Task;
 import org.example.taskmanagmentsystem.entity.User;
-import org.example.taskmanagmentsystem.entity.enums.Priority;
+import org.example.taskmanagmentsystem.entity.enums.Status;
+import org.example.taskmanagmentsystem.exeption_handing.IncorrectStatus;
+import org.example.taskmanagmentsystem.exeption_handing.NoAuthorWithThisId;
+import org.example.taskmanagmentsystem.exeption_handing.YouCantChangeOtherTask;
+import org.example.taskmanagmentsystem.exeption_handing.YouCantChangeStatusOfNotYourTask;
 import org.example.taskmanagmentsystem.repository.TaskRepository;
 import org.example.taskmanagmentsystem.repository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 
+@Log4j2
 @Service
 public class TaskService {
 
@@ -26,8 +37,11 @@ public class TaskService {
     }
 
     public Task createTask(TaskDTO taskDTO) {
-        User assignee=userRepository.findById(taskDTO.getAssigneeId()).orElseThrow(()->new UsernameNotFoundException("Вы ввели id не существующего юзера"));
-        User author=userRepository.findById(taskDTO.getAuthorId()).orElseThrow(()->new UsernameNotFoundException("Вы ввели id не существующего юзера"));
+        User assignee = userRepository.findById(taskDTO.getAssigneeId())
+                .orElseThrow(() -> new UsernameNotFoundException("Вы ввели id не существующего юзера"));
+        String currentName = SecurityContextHolder.getContext().getAuthentication().getName();
+        User author = userRepository.findByEmail(currentName).orElseThrow(() -> new UsernameNotFoundException("Текущий пользователь не найден"));
+        log.info("Сейчас текущее email пользователя" + author.getEmail());
         Task newtask = Task.builder()
                 .title(taskDTO.getTitle())
                 .description(taskDTO.getDescription())
@@ -39,33 +53,64 @@ public class TaskService {
         return taskRepository.save(newtask);
     }
 
-    public Task updateTask(Long id, Task taskDetails) {
-        Task task = taskRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Task not found"));
-        task.setTitle(taskDetails.getTitle());
-        task.setDescription(taskDetails.getDescription());
-        task.setStatus(taskDetails.getStatus());
-        task.setPriority(taskDetails.getPriority());
-        task.setAssignee(taskDetails.getAssignee());
-        return taskRepository.save(task);
+    public Task changeTaskStatusByAssignee(Long id, String  status) {
+        Status statusEnum;
+        try{
+           statusEnum =Status.valueOf(status);
+        }catch (Exception e){
+            throw new IncorrectStatus();
+        }
+        String currentName = SecurityContextHolder.getContext().getAuthentication().getName();
+       Task task=taskRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Задача не найдена"));
+
+       if (!task.getAssignee().getEmail().equals(currentName)){
+           throw new YouCantChangeStatusOfNotYourTask();
+       }
+       task.setStatus(statusEnum);
+      return taskRepository.save(task);
     }
 
-    public Task assignTask(Long taskId, Long userId) {
-        Task task = taskRepository.findById(taskId).orElseThrow(() -> new EntityNotFoundException("Task not found"));
-        User assignee = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("User not found"));
+    public Task updateTaskForAuthor(Long id, TaskDTO taskDTO) {
+        String currentName = SecurityContextHolder.getContext().getAuthentication().getName();
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Задача не найдена"));;
+        if (!task.getAuthor().getEmail().equals(currentName)) {
+            throw new YouCantChangeOtherTask();
+        }
+
+        User assignee = userRepository.findById(taskDTO.getAssigneeId())
+                .orElseThrow(() -> new UsernameNotFoundException("Вы ввели id не существующего юзера"));
+        task.setTitle(taskDTO.getTitle());
+        task.setDescription(taskDTO.getDescription());
+        task.setPriority(taskDTO.getPriority());
         task.setAssignee(assignee);
         return taskRepository.save(task);
     }
 
-    public void deleteTask(Long id) {
+    public void deleteTaskForAuthor(Long id) {
+        String currentName = SecurityContextHolder.getContext().getAuthentication().getName();
+        Task task = taskRepository.findById(id)
+                .orElseThrow(() -> new EntityNotFoundException("Задача не найдена"));
+        if (!task.getAuthor().getEmail().equals(currentName)) {
+            throw new YouCantChangeOtherTask();
+        }
         taskRepository.deleteById(id);
     }
 
-    public List<Task> getTasksForUser(Long userId) {
-        User user = userRepository.findById(userId).orElseThrow(() -> new EntityNotFoundException("User not found"));
-        return taskRepository.findByAssignee(user);
+    public Page<Task> getAuthorTasks(Long authorId, int page, int size, String sortBy) {
+        User author = userRepository.findById(authorId)
+                .orElseThrow(() -> new NoAuthorWithThisId());
+
+        Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
+        return taskRepository.findAllByAuthor(author, pageable);
     }
 
-    public List<Task> getAllTasks() {
-        return taskRepository.findAll();
+    public Page<Task> getAssigneeTasks(Long assigneeId, int page, int size, String sortBy) {
+        User assignee = userRepository.findById(assigneeId)
+                .orElseThrow(() -> new NoAuthorWithThisId());
+
+        Pageable pageable = PageRequest.of(page, size, Sort.by(sortBy));
+        return taskRepository.findAllByAssignee(assignee, pageable);
     }
+
 }
